Features/verify account

+++ src/redux/actions/role.action.ts
@@ -12,12 +12,11 @@ import {
 } from '@/redux/reducers/role.reducer';
 import { AppDispatch, RootState } from '@/redux/store';
 
-// import { PATH_NAMES } from '@/routes/pathnames';
-
 import { actionWithLoader } from '@/utils/app.utils';
-import { HIGHEST_LEVEL_DEFAULT_ROLES, ROLE_DEFAULT_LIMIT } from '@/utils/constants';
+import { DEFAULT_ROLES, HIGHEST_LEVEL_DEFAULT_ROLES, ROLE_DEFAULT_LIMIT } from '@/utils/constants';
 
-import { ILoadRolesInput, IRole, RoleInput, RolesForUserInput } from '@/types/role.type';
+import { ILoadRolesInput, IRole, IRoleInput, RolesForUserInput } from '@/types/role.type';
+import { PATH_NAMES } from '@/utils/pathnames';
 
 export const loadRoles = ({
   limit = ROLE_DEFAULT_LIMIT,
@@ -70,7 +69,7 @@ export const loadRoles = ({
  * @param values
  * @returns
  */
-export const createRole = (values: RoleInput): any => {
+export const createRole = (values: IRoleInput): any => {
   return actionWithLoader(async (dispatch: AppDispatch): Promise<void> => {
     // --------- update database --------- //
     const body = {
@@ -90,7 +89,7 @@ export const createRole = (values: RoleInput): any => {
   });
 };
 
-export const updateRole = (name: string, values: RoleInput): any => {
+export const updateRole = (name: string, values: IRoleInput): any => {
   return actionWithLoader(async (dispatch: AppDispatch): Promise<void> => {
     // --------- update database --------- //
     const body = {
@@ -110,10 +109,10 @@ export const updateRole = (name: string, values: RoleInput): any => {
   });
 };
 
-export const addRolesToUser = (userId: string, values: RolesForUserInput, currentRoles: IRole[]): any => {
+export const addRolesToUser = (userId: string, values: RolesForUserInput, currentRoles?: IRole[]): any => {
   return actionWithLoader(async (dispatch: AppDispatch): Promise<void> => {
     // --------- update database --------- //
-    const currentRoleNames = currentRoles.map((role: IRole): string => role.name);
+    const currentRoleNames = currentRoles?.map((role: IRole): string => role.name) || [];
     const allRoleNames = [...values.roles, ...currentRoleNames];
 
     const body = { userId, roleNames: allRoleNames };
@@ -169,10 +168,18 @@ export const destroyRole = (id: string): any => {
   return actionWithLoader(async (dispatch: AppDispatch): Promise<void> => {
     // update the database
     const role = await new Parse.Query(Parse.Role).equalTo('objectId', id).first();
-    // display success message
-    await role?.destroy();
+    if (!role) {
+      throw new Error(i18n.t('user:role.roleNotFound'));
+    }
 
-    dispatch(setMessageSlice(i18n.t('user:role.deleteRoleSuccessfully', { name: role?.getName() })));
+    if (DEFAULT_ROLES.includes(role.getName())) {
+      throw new Error(i18n.t('user:role.roleCannotBeDeleted'));
+    }
+
+    // display success message
+    await role.destroy();
+
+    dispatch(setMessageSlice(i18n.t('user:role.deleteRoleSuccessfully', { name: role.getName() })));
 
     // reload the list of deleted rides
     await dispatch(destroyRoleSlice(id));
@@ -188,7 +195,7 @@ export const onRolesEnter = (): any => {
   };
 };
 
-// --------------------------------------- //
-// ------------- redirection ------------- //
-// --------------------------------------- //
-// export const goToRoles = (): UpdateLocationAction<'push'> => push(PATH_NAMES.roles);
+// ---------------------------------------- //
+// -------------- redirection ------------- //
+// ---------------------------------------- //
+export const goToRoles = () => ({ to: PATH_NAMES.roles });
